Add explicit cast (gl_InstanceID is int), fixes errors on mesa compiler

+++ renderdoc/data/glsl/text.vert
@@ -39,7 +39,7 @@ void main(void)
                                   vec3( 1.0,  1.0, 0.5));
 
 	vec3 pos = verts[VERTEX_ID];
-	uint strindex = INSTANCE_ID;
+	uint strindex = uint(INSTANCE_ID);
 	
 	vec2 charPos = vec2(strindex + pos.x + general.TextPosition.x, pos.y + general.TextPosition.y);
